Features added:
1. Monit before movie deletion added.
2. Delete button added, instead of link.
3. Shadow text added to the Add movie form as
example, what to write in the input fields.
4. Check if year is a number, check if year is
not empty.
5. Title length changed to 2 (for movies like
"IT" or "Us").
6. "directed by" shows only if director
is not empty.
6. Polish "Tytuł" changed to "Title".

+++ ui/src/App.js
@@ -39,24 +39,21 @@ function App() {
         
       }, []);
 
-
       async function handleDeleteMovie(movie) {
+        const confirmed = window.confirm(`Are you sure you want to delete this movie "${movie.title}"?`);
+        if (!confirmed) {
+          return;
+        }
 
         const response = await fetch(`/movies/${movie.id}`, {
-    
-            method: 'DELETE',
-    
+          method: 'DELETE',
         });
-    
+
         if (response.ok) {
-    
-            const nextMovies = movies.filter(m => m !== movie);
-    
-            setMovies(nextMovies);
-    
+          const nextMovies = movies.filter(m => m !== movie);
+          setMovies(nextMovies);
         }
-    
-    }
+      }
 
 
     return (
